+++ nchuOJ-backend/nchuOJ-backend-model/src/main/java/com/jueye/nchuojbackendmodel/model/dto/question/JudgeConfig.java
@@ -10,13 +10,13 @@ public class JudgeConfig {
     /**
      *时间限制 ms
      */
-    private Long timeLimit;
+    private Integer timeLimit;
     /**
      * 内存限制 kb
      */
-    private Long memoryLimit;
+    private Integer memoryLimit;
     /**
      * 堆栈限制
      */
-    private Long stackLimit;
+    private Integer stackLimit;
 }
